Added 2-octet ASN support.
Cisco IOS XR/XE will send unmodified UPDATES, which will contain 2-octet ASN's if 4-octet is not negotiated.  https://tools.ietf.org/html/draft-ietf-grow-bmp-07#section-5 specifies that all UPDATES should use 4-octet ASN regardless of sent/received capabilities.  Most of the time 4-octet ASN's are used, so this is normally not an issue.  Legacy hardware/software still exists and therefore 2-octet ASN peering will be used.  This update attempts to auto-detect if 2-octet ASN parsing should be used by using the following logic:

Default to 4-octet and attempt to parse the ASN's (path), regardless of the negotiated peering capabilities as per draft-07.   If this fails, then check both sent/received capability to confirm 4-octet ASN support is not available.   If not 4-octet, attempt to parse using 2-octet.

AS4_PATH/aggregator is not currently implemented, but will be if this is really needed.