Fixed issue with 2-octet AS_PATH parsing.
Latest bmp draft includes a peer header A flag to indicate the AS PATH encoding
as either 2 or 4 octets.

Most implementations do not implement this yet. Previous implementations were supposed
to always use 4 octet regardless of peer capability.  Some implementations ignored
this and mirrored the UPDATES with 2-octet encoding.   The previous detection for this
had an issue when there were multiple segments.

The detection logic has changed to the following:

* If peer negotiated capability is four octet, then stick with four octet
  regardless (requires sent and recv to have 4-octet cap)

* If peer negotiated capability is two octet, then try to read/parse the
  complete AS path (all segments) using 4 octet. If the result is negative,
  set peer parsing using 2-octet.  This check is only needed once on first RM.