Allow bind IPs to be specified
Currently the logic binds to all interfaces, there are occasions where this is
not ideal, such as where there is a dedicated interface to an isolated network.

This adds 2 new base properties:

* listen_ipv4 - Takes an IPv4 address to bind to
* listen_ipv6 - Takes an IPv6 address to bind to

The default behaviour is still to bind to all interfaces, maintaining
compatibility.

Invalid bind addresses result in a throw of either 'ERROR: Cannot bind to IPv4
address and port' or 'ERROR: Cannot bind to IPv6 address and port'.