Fix use of unitialized variable
In the case that the `hash_type` is neither 1 or 2 (e.g., when it is 0), then the `hash_val` variable is used unitialized.
This change adds an extra `else` clause that uses the IP address for the hash (type 0 according to https://github.com/OpenBMP/openbmp/blob/master/Server/src/MsgBusInterface.hpp#L72) when `hash_type` is not 1 or 2.